Rename portforio component
